% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{.onLoad}
\alias{.onLoad}
\title{Executed on load of the library.}
\usage{
.onLoad(lib, pkg)
}
\arguments{
\item{lib}{The name of the library}

\item{pkg}{The name of the package}
}
\description{
Executed on load of the library.
}
