% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gar.R
\docType{package}
\name{gar}
\alias{gar}
\title{gar: R bindings for the gaze analysis C library libgac.}
\description{
The library allow to perform basic gaze analysis.
\itemize{
\item Sample filtering with moving average
\item Sample gap fill-in through linear interpolation (lerp)
\item Fixation detection with I-DT algorithm
\item Saccade detection with I-VT algorithm
}
}
