% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_add_aoi}
\alias{gar_add_aoi}
\title{Add an arbitrary AOI to the gaze analysis handler in order to enable the AOI
analysis. Use gar_create_aoi() to create the AOI.}
\usage{
gar_add_aoi(h, aoi)
}
\arguments{
\item{h}{A pointer to the gaze analysis handler, holding the filter parameters.}

\item{aoi}{A pointer to the AOI to add.}
}
\description{
Add an arbitrary AOI to the gaze analysis handler in order to enable the AOI
analysis. Use gar_create_aoi() to create the AOI.
}
\examples{
 h <- gar_create()
 x <- c( 0.5, 0.5, 0.6, 0.7, 0.8, 0.8, 0.7, 0.6 )
 y <- c( 0.4, 0.3, 0.2, 0.2, 0.3, 0.4, 0.5, 0.5 )
 df <- data.frame( x, y )
 aoi <- gar_create_aoi( df, "myAOI" )
 gar_add_aoi( h, aoi )
}
