% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_add_aoi_points}
\alias{gar_add_aoi_points}
\title{Add an AOI defined by points to the gaze analysis handler in order to enable
the AOI analysis.}
\usage{
gar_add_aoi_points(h, points, label = NULL)
}
\arguments{
\item{h}{A pointer to the gaze analysis handler.}

\item{points}{A data frame holding normalized 2d points.}

\item{label}{An optional label describing the AOI.}
}
\description{
Add an AOI defined by points to the gaze analysis handler in order to enable
the AOI analysis.
}
\examples{
 h <- gar_create()
 x <- c( 0.5, 0.5, 0.6, 0.7, 0.8, 0.8, 0.7, 0.6 )
 y <- c( 0.4, 0.3, 0.2, 0.2, 0.3, 0.4, 0.5, 0.5 )
 df <- data.frame( x, y )
 gar_add_aoi_points( h, df, "myAOI" )
}
