% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_create}
\alias{gar_create}
\title{Create a gaze analysis handler. If no parameter structure is provided
default values are used.}
\usage{
gar_create(params = NULL)
}
\arguments{
\item{params}{An optional filter parameter structure.}
}
\value{
A pointer to the allocated structure or NULL on failure.
}
\description{
Create a gaze analysis handler. If no parameter structure is provided
default values are used.
}
\examples{
 h <- gar_create()

 params <- gar_get_filter_parameter_default()
 params$gap$max_gap_length <- 0
 h <- gar_create( params )
}
