% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_create_aoi}
\alias{gar_create_aoi}
\title{Create an AOI structure given a list of points.}
\usage{
gar_create_aoi(points, label = NULL)
}
\arguments{
\item{points}{A data frame holding normalized 2d points.}

\item{label}{An optional label describing the AOI.}
}
\value{
A pointer to the allocated structure or NULL on failure.
}
\description{
Create an AOI structure given a list of points.
}
\examples{
 x <- c( 0.5, 0.5, 0.6, 0.7, 0.8, 0.8, 0.7, 0.6 )
 y <- c( 0.4, 0.3, 0.2, 0.2, 0.3, 0.4, 0.5, 0.5 )
 df <- data.frame( x, y )
 aoi <- gar_create_aoi( df, "myAOI" )
}
