% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_get_filter_parameter}
\alias{gar_get_filter_parameter}
\title{Get the current filter parameters.}
\usage{
gar_get_filter_parameter(h)
}
\arguments{
\item{h}{A pointer to the gaze analysis handler}
}
\value{
The filter parameters on success or NULL on failure
}
\description{
Get the current filter parameters.
}
\examples{
 h <- gar_create()
 params <- gar_get_filter_parameter( h )
}
