% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_get_filter_parameter_default}
\alias{gar_get_filter_parameter_default}
\title{Get the default filter parameter values.}
\usage{
gar_get_filter_parameter_default()
}
\value{
The default filter parameters on success or NULL on failure.
The parameters are structured as nested named lists:
\itemize{
\item \code{gap}: The gap filter parameter
\itemize{
\item \code{max_gap_length}: The maximal allowed gap length to be filled-in. Set
to zero to disable gap fill-in filter.
\item \code{sample_period}: The sample period to compute the number of required
fill-in samples.
}
\item \code{noise}: The noise filter parameter
\itemize{
\item \code{mid_idx}: The mid index of the moving average noise filter window.
This is used to compute the length of the window:
\code{window_length = mid_idx * 2 + 1}. Set to zero to disable noise filter.
}
\item \code{fixation}: The fixation parser parameters
\itemize{
\item \code{duration_threshold}: The duration threshold in milliseconds.
\item \code{dispersion_threshold}: The dispersion threshold in degrees.
}
\item \code{saccade}: The saccade parser parameters
\itemize{
\item \code{velocity_threshold}: The velocity threshold in degrees per second.
}
}
}
\description{
Get the default filter parameter values.
}
\examples{
 params <- gar_get_filter_parameter_default()
}
