% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_set_screen}
\alias{gar_set_screen}
\title{Configure the screen position in 3d space. This allows to compute 2d
gaze point coordinates.}
\usage{
gar_set_screen(
  h,
  top_left_x,
  top_left_y,
  top_left_z,
  top_right_x,
  top_right_y,
  top_right_z,
  bottom_left_x,
  bottom_left_y,
  bottom_left_z
)
}
\arguments{
\item{h}{A pointer to the gaze analysis handler.}

\item{top_left_x}{The x coordinate of the top left screen corner.}

\item{top_left_y}{The y coordinate of the top left screen corner.}

\item{top_left_z}{The z coordinate of the top left screen corner.}

\item{top_right_x}{The x coordinate of the top right screen corner.}

\item{top_right_y}{The y coordinate of the top right screen corner.}

\item{top_right_z}{The z coordinate of the top right screen corner.}

\item{bottom_left_x}{The x coordinate of the bottom left screen corner.}

\item{bottom_left_y}{The y coordinate of the bottom left screen corner.}

\item{bottom_left_z}{The z coordinate of the bottom left screen corner.}
}
\description{
Configure the screen position in 3d space. This allows to compute 2d
gaze point coordinates.
}
\examples{
 h <- gar_create()
 gar_set_screen( h, -298.64, 331.74, 113.91, 298.88, 331.74, 113.91, -298.64, 15.91, -1.05 )
}
