% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_get_filter_parameter_default}
\alias{gar_get_filter_parameter_default}
\title{Get the default filter parameter values.}
\usage{
gar_get_filter_parameter_default()
}
\value{
The default filter parameters on success or NULL on failure
}
\description{
Get the default filter parameter values.
}
\examples{
 params <- gar_get_filter_parameter_default()
}
