% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gar_parse}
\alias{gar_parse}
\title{Parse a set of input data for fixations and saccades.}
\usage{
gar_parse(h, px, py, pz, ox, oy, oz, timestamp)
}
\arguments{
\item{h}{A pointer to the gaze analysis handler, holding the filter parameters.}

\item{px}{A double vector of x cooridnates of the gaze point}

\item{py}{A double vector of y cooridnates of the gaze point}

\item{pz}{A double vector of z cooridnates of the gaze point}

\item{ox}{A double vector of x cooridnates of the gaze origin}

\item{oy}{A double vector of y cooridnates of the gaze origin}

\item{oz}{A double vector of z cooridnates of the gaze origin}

\item{timestamp}{A double vector of the relative timestamp in milliseconds}
}
\value{
The identified fixations and saccades as a named list:
 - fixations[]:
   - px: The x-coordinate of the average gaze point during the fixation
   - py: The y-coordinate of the average gaze point during the fixation
   - pz: The z-coordinate of the average gaze point during the fixation
   - duration: The fixation duration in milliseconds
   - timestamp: The relative timestamp of the first gaze point in the
     fixation in milliseconds
 - saccades[]:
   - startx: The x-coordinate of the first gaze point in the saccade
   - starty: The y-coordinate of the first gaze point in the saccade
   - startz: The z-coordinate of the first gaze point in the saccade
   - destx: The x-coordinate of the last gaze point in the saccade
   - desty: The y-coordinate of the last gaze point in the saccade
   - destz: The z-coordinate of the last gaze point in the saccade
   - duration: The saccade duration in milliseconds
   - timestamp: The relative timestamp of the first gaze point in the
     saccade in milliseconds
}
\description{
Parse a set of input data for fixations and saccades.
}
\examples{
 h <- gar_create()
 d <- read.csv( 'sample.csv', colClasses=c('numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric') )
 res <- gar_filter( h, d$px, d$py, d$pz, d$ox, d$oy, d$oz, d$timestamp )
}
